requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.1.0=gpu
absl-py=0.8.1=py37_0
asn1crypto=1.2.0=py37_0
astor=0.8.0=py37_0
attrs=19.3.0=pypi_0
backcall=0.1.0=py37_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2019.11.27=0
cairo=1.14.12=h8948797_3
certifi=2019.11.28=py37_0
cffi=1.13.2=py37h2e261b9_0
chardet=3.0.4=py37_1003
cloudpickle=1.2.2=py_0
conda=4.8.1=py37_0
conda-package-handling=1.6.0=py37h7b6447c_0
cryptography=2.8=py37h1ba5d50_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cupti=10.0.130=0
cycler=0.10.0=py37_0
cytoolz=0.10.1=py37h7b6447c_0
dask-core=2.9.1=py_0
dbus=1.13.12=h746ee38_0
decorator=4.4.1=py_0
dill=0.3.1.1=pypi_0
expat=2.2.6=he6710b0_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.9.1=h8a8886c_1
future=0.18.2=pypi_0
gast=0.2.2=py37_0
glib=2.63.1=h5a9c865_0
google-pasta=0.1.8=py_0
googleapis-common-protos=1.6.0=pypi_0
graphite2=1.3.13=h23475e2_0
grpcio=1.16.1=py37hf8bcb03_1
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.8.0=py37h989c5e5_3
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=h9c2bf20_1
idna=2.8=py37_0
imageio=2.6.1=py37_0
intel-openmp=2019.4=243
ipython=7.11.1=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
jasper=2.0.14=h07fcdf6_1
jedi=0.15.2=py37_0
jpeg=9b=h024ee3a_2
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.1.0=py37he6710b0_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libopencv=3.4.2=hb342d67_1
libopus=1.3=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.11.2=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=h439df22_0
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
markdown=3.1.1=py37_0
matplotlib=3.1.1=py37h5429711_0
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
ncurses=6.1=he6710b0_1
networkx=2.4=py_0
ninja=1.9.0=py37hfd86e86_0
numpy=1.17.4=py37hc1035e2_0
numpy-base=1.17.4=py37hde5b4d6_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
openssl=1.1.1d=h7b6447c_3
opt_einsum=3.1.0=py_0
parso=0.5.2=py_0
pcre=8.43=he6710b0_0
pexpect=4.7.0=py37_0
pickleshare=0.7.5=py37_0
pillow=6.2.1=py37h34e0f95_0
pip=19.3.1=py37_0
pixman=0.38.0=h7b6447c_0
promise=2.3=pypi_0
prompt_toolkit=3.0.2=py_0
protobuf=3.11.2=py37he6710b0_0
ptyprocess=0.6.0=py37_0
py-opencv=3.4.2=py37hb342d67_1
pycosat=0.6.3=py37h14c3975_0
pycparser=2.19=py37_0
pygments=2.5.2=py_0
pyopenssl=19.1.0=py37_0
pyparsing=2.4.6=py_0
pyqt=5.9.2=py37h05f1152_2
pysocks=1.7.1=py37_0
python=3.7.4=h265db76_1
python-dateutil=2.8.1=py_0
pytorch=1.3.1=py3.7_cuda10.0.130_cudnn7.6.3_0
pytz=2019.3=py_0
pywavelets=1.1.1=py37h7b6447c_0
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
requests=2.22.0=py37_1
ruamel_yaml=0.15.46=py37h14c3975_0
scikit-image=0.15.0=py37he6710b0_0
scipy=1.3.2=py37h7c811a0_0
setuptools=42.0.2=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.13.0=py37_0
sqlite=3.30.1=h7b6447c_0
tensorboard=2.0.0=pyhb38c66f_1
tensorflow=2.0.0=gpu_py37h768510d_0
tensorflow-base=2.0.0=gpu_py37h0ec5d1f_0
tensorflow-datasets=1.3.2=pypi_0
tensorflow-estimator=2.0.0=pyh2649769_0
tensorflow-gpu=2.0.0=h0d30ee6_0
tensorflow-hub=0.7.0=pypi_0
tensorflow-metadata=0.21.0=pypi_0
termcolor=1.1.0=py37_1
tk=8.6.8=hbc83047_0
toolz=0.10.0=py_0
torchvision=0.4.2=py37_cu100
tornado=6.0.3=py37h7b6447c_0
tqdm=4.40.0=py_0
traitlets=4.3.3=py37_0
urllib3=1.25.7=py37_0
wcwidth=0.1.7=py37_0
werkzeug=0.16.0=py_0
wheel=0.33.6=py37_0
wrapt=1.11.2=py37h7b6447c_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
